pyproject.toml
[project]
name = "qprompts"
version = "0.1.0"
description = "An ergonomic library for creating, managing, and versioning prompts for Large Language Models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.4",
    "prompty>=0.1.24",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
]
